feat: Implement product CSV import with bulk creation and add API key authentication middleware for protected routes.

+++ backend/src/features/products/product-controller.ts
@@ -1,6 +1,36 @@
 import { Request, Response } from "express";
 import prisma from "../../core/config/prisma.js";
+import fs from "fs";
 import { createProductSchema, updateProductSchema } from "./product.schema.js";
+import { parseProductsFromCsv } from "./services/csv-service.ts";
+import { bulkCreateProducts } from "./services/product-service.ts";
+
+export const getProductById = async (req: Request, res: Response) => {
+  try {
+    // get product id
+    const { id } = req.params;
+
+    // check if id is not null
+    if (!id) {
+      return res.status(400).json({ success: false, error: "Missing product id" });
+    }
+
+    // get product from database
+    const product = await prisma.product.findUnique({
+      where: { id: Number(id) }
+    });
+
+    // return product
+    return res.status(200).json({ success: true, data: product });
+  } catch (error) {
+    if (error.code === 'P2025') {
+      return res.status(404).json({ success: false, error: "Cannot find product with id" });
+    }
+    console.error("Error during product retrieval:");
+    console.error(error);
+    return res.status(500).json({ success: false, error: "Internal server error" });
+  }
+}
 
 // TODO: add ZOD validation for page and limit
 export const getAllProducts = async (req: Request, res: Response) => {
@@ -100,8 +130,13 @@ export const deleteProduct = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    // check if id is not null
+    if (!id) {
+      return res.status(400).json({ success: false, error: "Missing product id" });
+    }
+
     const product = await prisma.product.delete({
-      where: { id }
+      where: { id: Number(id) }
     });
 
     return res.status(200).json({ success: true, data: product });
@@ -113,5 +148,38 @@ export const deleteProduct = async (req: Request, res: Response) => {
     console.error("Error during product deletion:");
     console.error(error);
     return res.status(500).json({ success: false, error: "Internal server error" });
+  }
+}
+
+export const importProductsFromCsv = async (req: Request, res: Response) => {
+  // check if file is uploaded
+  if (!req.file) {
+    return res.status(400).json({ success: false, error: "Nessun file CSV caricato" });
+  }
+
+  // get file path
+  const filePath = req.file.path;
+
+  try {
+    // parse data
+    const data = await parseProductsFromCsv(filePath);
+
+    // create products in db
+    const result = await bulkCreateProducts(data);
+
+    // return result
+    return res.status(201).json({ 
+      message: "Importato!", 
+      count: result.count 
+    });
+  } catch (error) {
+    console.error("Error during product import:");
+    console.error(error);
+    return res.status(500).json({ success: false, error: "Internal server error" });
+  } finally {
+    // delete file
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
   }
 }
